requirements.txt
# =============================================================================
# BizFlow AI Service — Python dependencies
# =============================================================================

# --- Web framework -----------------------------------------------------------
fastapi==0.115.6
uvicorn[standard]==0.32.1

# --- Data validation & settings ----------------------------------------------
pydantic==2.10.3
pydantic-settings==2.7.0

# --- Database ----------------------------------------------------------------
sqlalchemy==2.0.36
pymysql==1.1.1

# --- OpenAI (LLM + Whisper STT + Vision OCR) ---------------------------------
openai==1.58.1

# --- Google Cloud STT (primary STT provider) ---------------------------------
google-cloud-speech==2.28.0

# --- Vector store & embeddings (RAG for product catalog) ---------------------
chromadb==0.5.23
sentence-transformers==3.3.1

# --- Data processing (forecasting, reorder, product insights) ----------------
pandas==2.2.3
numpy==2.1.3

# --- File upload handling ----------------------------------------------------
python-multipart==0.0.20

# --- Async HTTP (internal health checks / optional) -------------------------
httpx==0.28.1

# --- Database migrations -----------------------------------------------------
alembic==1.14.1
